% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_history_export.R
\name{gg_history_export}
\alias{gg_history_export}
\title{gg_history_export}
\usage{
gg_history_export(
  to_powerpoint = TRUE,
  to_zip = TRUE,
  ppt_name = NULL,
  zip_name = NULL,
  export_folder = NULL,
  create_export_folder = FALSE,
  stoprecording = FALSE,
  regex_pattern = "\\\\d{4}_\\\\d{2}_\\\\d{2}_\\\\d{2}_\\\\d{2}_\\\\d{2}\\\\.\\\\d+"
)
}
\arguments{
\item{to_powerpoint}{Export recorded ggplots to a powerpoint presentation. Images will be converted to png.}

\item{to_zip}{Export recorded ggplots to a zip folder.}

\item{ppt_name}{Name of the PowerPoint file.}

\item{zip_name}{Name of the zip file.}

\item{export_folder}{Destination folder for exports.}

\item{create_export_folder}{Create the export folder (and applicable path) if it doesn't exist.}

\item{stoprecording}{Stop camcorder::gg_record from recording any additional plots after gg_history_export is finished exporting.}

\item{regex_pattern}{Optional regex filter for plots if change to file name recognition required.}
}
\description{
Adaptation of camcorder::gg_playback that provides options to export in additional formats. This functions requires the use of camcorder::gg_record to generate the images for export.
}
